add a postcss7 compat version
