+++ plugins/nesting/index.postcss7.js
@@ -1,0 +1,4 @@
+let postcss = require('postcss')
+let nesting = require('./plugin')
+
+module.exports = postcss.plugin('tailwindcss/nesting', nesting)
